\name{AnalysisTool.run}
\alias{AnalysisTool.run}
\title{Analysis Tool}
\usage{
AnalysisTool.run()
}
\description{
Run the shiny app "Analysis Tool" used to visualize the efficiency of clustering algorithms and the effects of the values of parameters on the variations of the F-score.
}
\examples{
%YOU MIGHT HAVE TO LOAD ANOTHER R SESSION BEFORE USING THE TOOL
library(AnalysisTool)
AnalysisTool.run()
}
